Extends jwt validation with Roles;
adds a decorator for checking the role set on the
token.

This allows us to set arbitrary access requirements on API routes.
For example: some UPDATE or CREATE actions should be reserved to an
admin only role, whereas other UPDATES may be done by a data-entry
role.

We default to "client" as the least-privileged role with READ access
only.